renamed mock_classes to be more specific
