+++ squidDrone/tests/mock_libraries/propulsion_mock_dependencies.hpp
